projects/sentiment/pyproject.toml
[project]
name = "sentiment"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
    "datasets>=2.0",
]
